MDS - Stock prices cache (#4)
* Connecting to brapi.dev

* Historical prices endpoint

* Improving brapi.dev health check

* Fix internal class unit testing

* Fix seq and rabbitmq ACA volume mounting

* Streaming and formatting the brapi quotes

* Global response formatter

* Global error handler

* Fix OutOfRangeException bug

* Adding MongoDB connection

* Fix brapi quote casting

* Creating Price mongo collection

* Caching price data in MongoDB

* Unit tests

* Fixing unit tests

* Post-merge fix

* MDS - Stock search (#5)

* Stock Searching

* Idempotency on stock caching

* Segregating Price and Stock repositories

* Idempotency on prices appending

* Stock cache TTL

* Pagination

* Unit tests

* Managed identity fix

* Fix in-memory stock filtering when Sector null

+++ src/MarketData/Infra/Dtos/StockQuote.cs
@@ -5,15 +5,15 @@ public sealed class StockQuote
     // Date formatted as Unix Timestamp (e.g., 1756126800)
     public required long Date { get; init; }
 
-    public required double? Open { get; init; }
+    public required decimal? Open { get; init; }
 
-    public required double? High { get; init; }
+    public required decimal? High { get; init; }
 
-    public required double? Low { get; init; }
+    public required decimal? Low { get; init; }
 
-    public required double? Close { get; init; }
+    public required decimal? Close { get; init; }
 
     public required ulong? Volume { get; init; }
 
-    public required double? AdjustedClose { get; init; }
+    public required decimal? AdjustedClose { get; init; }
 }
