+++ src/Shared/Contracts/MarketData/GetStockQuotes/Quote.cs
@@ -2,11 +2,11 @@ namespace Kairos.Shared.Contracts.MarketData.GetStockQuotes;
 
 public sealed record Quote(
     DateTime Date,
-    double Close,
-    double CloseWithEvents
+    decimal Close,
+    decimal CloseWithEvents
 )
 {
-    public Quote(long unixTimeSeconds, double close, double adjustedClose) : this(
+    public Quote(long unixTimeSeconds, decimal close, decimal adjustedClose) : this(
         DateTimeOffset.FromUnixTimeSeconds(unixTimeSeconds).Date,
         close,
         adjustedClose
